pyproject.toml
[project]
name = "ha-kubernetes-integration"
version = "0.1.0"
description = "Home Assistant custom integration for Kubernetes"
requires-python = ">=3.12"
dependencies = [
    "kubernetes>=31.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-homeassistant-custom-component>=0.13",
    "pre-commit>=4.0",
    "ruff>=0.11",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
